Start adding comments

+++ src/NImpeller/ImpellerColor.cs
@@ -2,7 +2,7 @@ namespace NImpeller;
 
 public partial record struct ImpellerColor
 {
-    public static ImpellerColor FromArgb(int a, int r, int g, int b) =>
+    public static ImpellerColor FromArgb(float a, float r, float g, float b) =>
         new()
         {
             Alpha = a / 255.0f,
@@ -11,6 +11,6 @@ public partial record struct ImpellerColor
             Blue = b / 255.0f
         };
     
-    public static ImpellerColor FromRgb(int r, int g, int b) =>
+    public static ImpellerColor FromRgb(float r, float g, float b) =>
         FromArgb(255, r, g, b);
 }
